pyproject.toml
[project]
name = "whatsurname"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dspy>=3.1.2",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "fire>=0.5.0",
    "textual>=0.47.0",
]

[project.scripts]
whatsurname = "whatsurname.main:main"

[tool.uv]
package = true
